+++ Plugin.cs
@@ -1,40 +1,32 @@
 using BepInEx;
 using BepInEx.Logging;
+using SigilADay_julianperge.lib;
+using UnityEngine;
 
-namespace SigilADay
+// using UnityEngine;
+
+namespace SigilADay_julianperge
 {
-  [BepInPlugin(PluginGuid, PluginName, PluginVersion)]
-  [BepInDependency("cyantist.inscryption.api", BepInDependency.DependencyFlags.HardDependency)]
-  public partial class Plugin : BaseUnityPlugin
-  {
-    private const string PluginGuid = "julianperge.inscryption.sigiladay";
-    private const string PluginName = "SigilADay_julianperge";
-    private const string PluginVersion = "1.0.0.0";
+	[BepInPlugin(PluginGuid, PluginName, PluginVersion)]
+	[BepInDependency("cyantist.inscryption.api", BepInDependency.DependencyFlags.HardDependency)]
+	public partial class Plugin : BaseUnityPlugin
+	{
+		public const string PluginGuid = "julianperge.inscryption.sigiladay";
+		private const string PluginName = "SigilADay_julianperge";
+		private const string PluginVersion = "1.1.0";
+		public static AssetBundle ArtworkAssets;
 
-    internal static ManualLogSource Log;
+		internal static ManualLogSource Log;
 
-    private void Awake()
-    {
-      Log = base.Logger;
-
-      AddExcavator();
-      AddNest();
-      AddSplit();
-
-      // AddBloodGuzzler();
-      // AddLeech();
-      // AddRegen1();
-      // AddRegen2();
-      // AddRegen3();
-      // AddRegenFull();
-      // AddPoisonous();
-      // AddThickShell();
-      // AddBonePicker();
-      // AddNutritious();
-      // AddTransient();
-      //AddSilence();
-
-      // ChangeRingworm();
-    }
-  }
+		private void Awake()
+		{
+			Log = base.Logger;
+			
+			// ArtworkAssets = AssetBundle.LoadFromFile(SigilUtils.GetFullPathOfFile("artwork_sigils"));
+			
+			AddExcavator();
+			// AddNest();;
+			// AddSplit();
+		}
+	}
 }