+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## v1.1.0 
-- Corrected logic for checking for if PlayerSlots are full
-- Created `Flight of the Valkyrie` sigil.
--- When the creature is played, fill all open slots on your side of the field with 1/1 Norse Warriors
+- Updated to use API v1.12.1
+- Corrected logic for checking for if PlayerSlots are full in `Split` ability.
+- Created `Flight of the Valkyrie` ability.
+  * When the creature is played, fill all open slots on your side of the field with 1/1 Norse Warriors
+  * Norse Warriors are defined as 1/1 normal cards. 
+- Added Resource image loading so that users no longer have to copy/paste the artwork folder.